use Float32Array instead of Int32Array

+++ packages/vector-packer/src/pack/util/array.js
@@ -51,5 +51,5 @@ export function getPosArrayType(max) {
     max = Math.abs(max);
     if (max < 128) return Int8Array;
     if (max < 65536 / 2) return Int16Array;
-    return Int32Array;
+    return Float32Array;
 }
