+++ packages/vt/src/common/Util.js
@@ -173,7 +173,8 @@ export function getPosArrayType(max) {
     max = Math.abs(max);
     if (max < 128) return Int8Array;
     if (max < 65536 / 2) return Int16Array;
-    return Int32Array;
+    //TODO 这里不能用Int32Array，可能是regl的bug
+    return Float32Array;
 }
 
 export function compileStyle(styles) {
